fix unsigned int being pushed below 0 and resulting in a biiig loop

+++ include/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: estettle <estettle@student.42lausanne.ch>  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/04 17:46:51 by estettle          #+#    #+#             */
-/*   Updated: 2024/11/19 16:10:28 by estettle         ###   ########.fr       */
+/*   Updated: 2024/11/19 17:39:51 by estettle         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -62,7 +62,7 @@ t_slab		*slab_last(t_slab *lst);
 void		slab_add_back(t_slab **lst, t_slab *new);
 void		slab_add_front(t_slab **lst, t_slab *new);
 void		stack_clear(t_slab **lst);
-uint16_t	slab_count(t_slab *lst);
+int32_t		slab_count(t_slab *lst);
 
 // Util functions
 void		ft_kill(t_slab **stack1, t_slab **stack2, int error_code);
